% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{binding_lapply}
\alias{binding_lapply}
\title{binding_lapply}
\usage{
binding_lapply(x, characteristics)
}
\arguments{
\item{x}{Element to bind}

\item{characteristics}{Characteristics to filter for}
}
\description{
binding_lapply
}
