% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-misc.R
\name{catalogue_jobs}
\alias{catalogue_jobs}
\title{Job catalogue}
\usage{
catalogue_jobs(
  sortcriterion = c("type", "status", "time"),
  pagelength = 100,
  language = "en"
)
}
\arguments{
\item{sortcriterion}{Should the results be sorted by the code or the content of the object(s)?}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of processing jobs (e.g. very large table retrievals)
}
\examples{
\dontrun{
catalogue_modifieddata(date = format(Sys.Date() - 30, "\%d.\%m.\%Y"))
}
}
