% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-misc.R
\name{catalogue_modifieddata}
\alias{catalogue_modifieddata}
\title{Modified data catalogue}
\usage{
catalogue_modifieddata(
  selection = NULL,
  date = "01.01.1970",
  pagelength = 100,
  language = "en"
)
}
\arguments{
\item{selection}{String to filter for the code or content of the object(s) to
be returned. Use of wildcard (*) possible.}

\item{date}{List data that have been made available or updated since this date
("dd.mm.yyyy")}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of tables according to selection
}
\examples{
\dontrun{
catalogue_modifieddata(date = format(Sys.Date() - 30, "\%d.\%m.\%Y"))
}
}
