% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-statistics.R
\name{catalogue_statistics}
\alias{catalogue_statistics}
\title{Statistic catalogue}
\usage{
catalogue_statistics(
  selection,
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  pagelength = 100,
  language = "en"
)
}
\arguments{
\item{selection}{String to filter for the code or content of the object(s) to
be returned. Use of wildcard (*) possible.}

\item{searchcriterion}{Does \code{selection} refer to the code or the content of the object(s)?}

\item{sortcriterion}{Should the results be sorted by the code or the content of the object(s)?}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of statistics according to selection
}
\examples{
\dontrun{
catalogue_statistics("124*")
}
}
