% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue_tables}
\alias{catalogue_tables}
\title{Table catalogue}
\usage{
catalogue_tables(
  selection,
  area = c("free", "user", "all"),
  pagelength = 100,
  language = "en"
)
}
\arguments{
\item{selection}{Code of the table(s) to be returned. Use of wildcard (*) possible.}

\item{area}{Area in which the object is stored:
\itemize{
\item "free" = objects in the public catalogue
\item "user" = objects saved to "Meine Tabellen"
}}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieve list of tables according to selection
}
\examples{
\dontrun{
catalogue_tables("124*")
}
}
