% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-variables.R
\name{catalogue_tables_by_variable}
\alias{catalogue_tables_by_variable}
\title{Catalogue of tables related to a variable}
\usage{
catalogue_tables_by_variable(
  name,
  selection = NULL,
  area = c("free", "user", "all"),
  pagelength = 100,
  language = "en",
  genesis = getOption("genesis")
)
}
\arguments{
\item{name}{Code of the variable (e.g. retrieved using \link{catalogue_variables})}

\item{selection}{String to filter for the code or content of the object(s) to
be returned. Use of wildcard (*) possible.}

\item{area}{Area in which the object is stored:
\itemize{
\item "free" = objects in the public catalogue
\item "user" = objects saved to "Meine Tabellen"
}}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}

\item{genesis}{Which GENESIS database should be used?

One of: "destatis", "regionalstatistik", "bildungsmonitoring", "bayern", "nrw", "sachsen-anhalt".

A default value can also be set to the global option \code{genesis} via
\code{\link[=options]{options()}}, e.g. \code{options(genesis = "destatis")}, so that \code{genesis}
does not have to be specified in the function call.}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of tables based on variable name
}
\examples{
\dontrun{
options(genesis = "destatis")
catalogue_tables_by_variable("KREISE")
}
}
