% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-misc.R
\name{catalogue_terms}
\alias{catalogue_terms}
\title{Term catalogue}
\usage{
catalogue_terms(selection, pagelength = 100, language = "en")
}
\arguments{
\item{selection}{String to filter for the code or content of the object(s) to
be returned. Use of wildcard (*) possible.}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of search terms according to selection
}
\examples{
\dontrun{
catalogue_terms("schu*", language = "de")
}
}
