% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_authentication.R
\name{check_credential_list_standalone}
\alias{check_credential_list_standalone}
\title{Function to check if credential list provided is valid}
\usage{
check_credential_list_standalone(credential_list, database)
}
\arguments{
\item{credential_list}{Credential list to be checked}

\item{database}{Database that is used for authentication}
}
\description{
Function to check if credential list provided is valid
}
