% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evas_list}
\alias{evas_list}
\title{List of EVAS codes}
\format{
\subsection{\code{evas_list}}{

A data frame with 1132 rows and 3 columns:
\describe{
\item{EVAS}{EVAS code}
\item{Beschreibung}{Details on the EVAS code}
\item{Titel}{Alternative desription of EVAS code contents}
}
}
}
\source{
\url{https://www.destatis.de/DE/Methoden/Revisionen/Glossar/EVAS.html}
}
\usage{
evas_list
}
\description{
List of EVAS codes
}
\keyword{datasets}
