% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evas_list_long_20220724}
\alias{evas_list_long_20220724}
\title{List of EVAS Codes}
\format{
\subsection{\code{evas_list_long_20220724}}{

A data frame with 1,097 rows and 3 columns:
\describe{
\item{EVAS}{EVAS code}
\item{Beschreibung}{Details on the EVAS code}
\item{Titel}{Alternative description of EVAS code contents}
...
}
}
}
\source{
\url{https://www.destatis.de/DE/Methoden/Revisionen/Glossar/EVAS.html}
}
\usage{
evas_list_long_20220724
}
\description{
List of EVAS Codes
}
\keyword{datasets}
