% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_}
\alias{find_}
\alias{find_tables}
\alias{find_statistics}
\alias{find_variables}
\alias{find_cubes}
\alias{find_timeseries}
\title{Find tables, statistics, variables, data cubes, or time series}
\usage{
find_(
  term,
  category = c("tables", "statistics", "variables", "cubes", "time-series"),
  pagelength = 100,
  language = "en"
)

find_tables(...)

find_statistics(...)

find_variables(...)

find_cubes(...)

find_timeseries(...)
}
\arguments{
\item{term}{Term to search for}

\item{category}{Category to be searched}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}

\item{...}{
  Arguments passed on to \code{\link[=find_]{find_}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieve lists of objects for a search term
(tables, statistics, characteristics, data cubes or time series)
}
\examples{
\dontrun{
find_("Abfall", category = "statistics", pagelength = 6, language = "de")
find_tables("Abfall", language = "de")
}
}
