% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{find_token}
\alias{find_token}
\title{find_token}
\usage{
find_token(input, error.input, text, sub_category)
}
\arguments{
\item{input}{Input to test result structure}

\item{error.input}{error.ignore TRUE or FALSE}

\item{text}{verbose TRUE or FALSE}

\item{sub_category}{sub_category character string}
}
\description{
find_token
}
