% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gen_api}
\alias{gen_api}
\title{Low-level function to interact with the Destatis GENESIS API}
\usage{
gen_api(method, ...)
}
\description{
Low-level function to interact with the Destatis GENESIS API
}
\examples{
gen_api("helloworld/whoami") |>
  httr2::resp_body_json()
}
