% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_auth.R
\name{gen_auth_ask}
\alias{gen_auth_ask}
\title{gen_auth_ask}
\usage{
gen_auth_ask(credential_type)
}
\arguments{
\item{credential_type}{Character string. Type of credential to ask for}
}
\value{
The user response
}
\description{
gen_auth_ask
}
