% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_auth.R
\name{gen_auth_get}
\alias{gen_auth_get}
\title{Retrieve database credentials}
\usage{
gen_auth_get(database = NULL)
}
\arguments{
\item{database}{Character string. The database to get the credentials for ('all', 'genesis', 'zensus' and 'regio').}
}
\value{
Credentials for the database(s) chosen by the user
}
\description{
Function to retrieve the credentials stored via \code{gen_auth_save()}
}
\examples{
\dontrun{
gen_auth_get("all")
}

}
