% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_auth.R
\name{gen_auth_path}
\alias{gen_auth_path}
\title{Helper function for authentication purposes}
\usage{
gen_auth_path(...)
}
\arguments{
\item{...}{Optional arguments for file.path()}
}
\value{
A file path for the storage of config files
}
\description{
Helper function for authentication purposes
}
