% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cube.R
\name{gen_cube}
\alias{gen_cube}
\title{Download Data Cube As Data.Frame}
\usage{
gen_cube(
  name,
  database = c("genesis", "regio", "nrw", "bildung"),
  credential_list = NULL,
  area = c("public", "user"),
  values = TRUE,
  metadata = TRUE,
  additionals = FALSE,
  startyear = NULL,
  endyear = NULL,
  timeslices = NULL,
  contents = NULL,
  regionalvariable = NULL,
  regionalkey = NULL,
  classifyingvariable1 = NULL,
  classifyingkey1 = NULL,
  classifyingvariable2 = NULL,
  classifyingkey2 = NULL,
  classifyingvariable3 = NULL,
  classifyingkey3 = NULL,
  stand = NULL,
  language = Sys.getenv("RESTATIS_LANG"),
  ...
)
}
\arguments{
\item{name}{Character string for a cube object.}

\item{database}{Character string. Indicator if the GENESIS ('genesis'), regionalstatistik.de ('regio'), landesdatenbank.nrw.de ('nrw') or bildungsmonitoring.de ('bildung') database is called.}

\item{credential_list}{A list containing the credentials for the databases to be accessed. If 'NULL' (default), the function will use the stored credentials from \code{gen_auth_get()}.}

\item{area}{Character string. The area in which the cube is stored.
Possible values:
\itemize{
\item \code{"public"}: cube in the public catalogue
\item \code{"user"}: cube in the user's account
\item \code{"all"}: both of the above
}}

\item{values}{Boolean. Should values be included?}

\item{metadata}{Boolean. Should metadata be included?}

\item{additionals}{Boolean. Should additional metadata be included?}

\item{startyear}{Four-digit integer. Only retrieve data from this year onward.
Note: Starting with {restatis} v0.4.2, no default values for the parameters ‘startyear’ and ‘endyear’ will be submitted anymore
(formerly 1900 to 2100 to fetch all available values). Thus, the resulting cube might contain a shorter time span than expected.
Instead, we encourage users to explicitly set the desired time span. Doing so, we follow the best practice to save API resources.
The new default of ‘NULL’ behaves as follows: Each database will always send data, as each cube is preconfigured with a standard time
span that varies for each statistic and each cube. If the user desires different time spans, she has to specify the values accordingly.}

\item{endyear}{Four-digit integer. Only retrieve data up to this year.
Note: Starting with {restatis} v0.4.2, no default values for the parameters ‘startyear’ and ‘endyear’ will be submitted anymore
(formerly 1900 to 2100 to fetch all available values). Thus, the resulting cube might contain a shorter time span than expected.
Instead, we encourage users to explicitly set the desired time span. Doing so, we follow the best practice to save API resources.
The new default of ‘NULL’ behaves as follows: Each database will always send data, as each cube is preconfigured with a standard time
span that varies for each statistic and each cube. If the user desires different time spans, she has to specify the values accordingly.}

\item{timeslices}{Integer. Number of timeslices (cumulative to \code{startyear} or \code{endyear}).}

\item{contents}{Character string. Names of required statistical specifications.}

\item{regionalvariable}{Character string. Code of the regional variable whose value
is specified in \code{regionalkey} to filter the results.}

\item{regionalkey}{Character string. One or more regional keys. Multiple values can be
supplied as a character vector or as a single string, with the regional keys
separated by commas. Use of wildcard (\code{*}) allowed.}

\item{classifyingvariable1}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey1} is applied.}

\item{classifyingkey1}{Character string. One or more values of a subject classification
(e.g. \code{"WZ93012"}). Applied to \code{classifyingvariable1}. Multiple keys can be
supplied as a character vector or as a single string, with the keys separated by
commas. Use of wildcard (\code{*}) allowed.}

\item{classifyingvariable2}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey2} is applied.}

\item{classifyingkey2}{Character string. One or more values of a subject classification.
Applied to \code{classifyingvariable2}. Multiple keys can be supplied as a character
vector or as a single string, with the keys separated by commas. Use of wildcard
(\code{*}) allowed.}

\item{classifyingvariable3}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey3} is applied.}

\item{classifyingkey3}{Character string. One or more values of a subject classification.
Applied to \code{classifyingvariable3}. Multiple keys can be supplied as a character
vector or as a single string, with the keys separated by commas. Use of wildcard
(\code{*}) allowed.}

\item{stand}{Character string, format \code{"DD.MM.YYYY"}. Only retrieve data updated
after this date.}

\item{language}{Character string. Search terms, returned messages and data descriptions
in German (\code{"de"}) or English (\code{"en"}).}

\item{...}{Additional parameters for the API call (see respective API documentation).
A valid specification of these will not be checked by the function, so wrongful
specification may lead to errors.}
}
\value{
A \link[tibble:tibble]{tibble}. Non-data contents of the data cube object are saved in
the \code{metadata} \link[base:attr]{attribute} of the data frame.
}
\description{
Download a cube as a data.frame from any supported database. Works not on all databases included in the package.
}
\examples{
\dontrun{
gen_cube("47414BJ002")
}

}
