% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_list_results.R
\name{gen_list_results}
\alias{gen_list_results}
\title{gen_list_results}
\usage{
gen_list_results(
  database = c("genesis", "zensus", "regio"),
  area = c("all", "public", "user"),
  ...
)
}
\arguments{
\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus') or regionalstatistik.de ('regio') database is called. Only one database can be addressed per function call. Default option is 'genesis'.}

\item{area}{Character string. Indicator from which area of the database the results are called. In general, 'all' is the appropriate solution. Default option is 'all'. Not used for 'statistics'.}

\item{...}{Additional parameters for the API call. These parameters are only affecting the call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list of all current results connected to the given user.
}
\description{
Function to list all current results connected to the given user in the GENESIS, Zensus 2022 or regionalstatistik.de database. Important note: For this function it is also possible to use \code{selection} parameter, making it possible to filter the results based on the 'code' of the object. For more details see \code{vignette("additional_parameter")}.
}
