% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_logincheck.R
\name{gen_logincheck}
\alias{gen_logincheck}
\title{gen_logincheck}
\usage{
gen_logincheck(database, verbose = FALSE)
}
\arguments{
\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus') or regionalstatistik.de ('regio') database is called. Only one database can be addressed per function call. Default option is 'genesis'.}

\item{verbose}{Boolean. In case of success, should a message be printed? Defaults to FALSE.}
}
\value{
Leads to an informative error message if the login check failed and returns FALSE invisibly. Invisibly returns TRUE otherwise.
}
\description{
Function to check if a login is possible for a certain database.
}
\examples{
\dontrun{
gen_logincheck("zensus")
}

}
