% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meta_data.R
\name{gen_meta_data}
\alias{gen_meta_data}
\title{gen_meta_data}
\usage{
gen_meta_data(
  code = NULL,
  category = c("Cube", "Statistic", "Table", "Variable", "Value"),
  error.ignore = FALSE,
  ...
)
}
\arguments{
\item{category}{a string. Specific object-types: 'Cube', 'Statistic', "Table", "Variable" and 'Value'. The function needs a specified object type.}

\item{error.ignore}{a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.}

\item{...}{Additional parameter of the Genesis API call. These parameters are only affecting the Genesis API call itself, no further processing.}

\item{codea}{string with a maximum length of 15 characters. Code from a Genesis-Object. Only one code per iteration.}
}
\value{
A list with all recalled elements from Genesis. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Search For Meta-Information For All Types Of Objects
}
\examples{
\dontrun{
# Find meta-information of the table with the code "11111"
object <- gen_meta_data(code = "11111", category = "Table")
}

}
