% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meta_data.R
\name{gen_meta_data_tab}
\alias{gen_meta_data_tab}
\title{gen_meta_data_tab: Meta-Information For Table}
\usage{
gen_meta_data_tab(code = NULL, error.ignore = FALSE, ...)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Genesis-Object. Only one code per iteration.}

\item{error.ignore}{a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.}

\item{...}{Additional parameter of the Genesis API call. These parameters are only affecting the Genesis API call itself, no further processing.}
}
\value{
A list with all recalled elements from Genesis. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to search for meta-information for a specific table.
}
\examples{
\dontrun{
# Find meta-information of the table with the code "11111"
object <- gen_meta_data_tab(code = "11111")
}
}
