% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meta_data.R
\name{gen_metadata_stats}
\alias{gen_metadata_stats}
\title{Search for Meta-Information for a Statistic}
\usage{
gen_metadata_stats(code = NULL, error.ignore = FALSE, ...)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a GENESIS-Object. Only one code per iteration.}

\item{error.ignore}{a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.}

\item{...}{Additional parameter of the GENESIS API call. These parameters are only affecting the GENESIS API call itself, no further processing.}
}
\value{
A list with all recalled elements from GENESIS. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to search for meta-information for a specific statistic.
}
\examples{
\dontrun{
# Find meta-information of the statistic with the code "12411"
object <- gen_metadata_stats(code = "12411")
}

}
