% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meta_data.R
\name{gen_metadata_var}
\alias{gen_metadata_var}
\title{gen_metadata_var}
\usage{
gen_metadata_var(code = NULL, error.ignore = FALSE, ...)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Genesis-Object. Only one code per iteration. "*"-Notation is possible.}

\item{error.ignore}{a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.}

\item{...}{Additional parameters for the Genesis API call. These parameters are only affecting the Genesis call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list with all recalled elements from Genesis. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to search for meta-information for a specific variable.
}
\examples{
\dontrun{
# Find meta-information of the variable with the code "FAMSTD"
object <- gen_metadata_var(code = "FAMSTD")
}

}
