% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_var2-val2.R
\name{gen_search_vars}
\alias{gen_search_vars}
\title{Search For Variables}
\usage{
gen_search_vars(
  code = NULL,
  database = c("all", "genesis", "zensus", "regio", "bayern", "nrw", "bildung", "st"),
  area = c("all", "public", "user"),
  sortcriterion = c("code", "content"),
  credential_list = NULL,
  pagelength = 500,
  error.ignore = FALSE,
  verbose = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{code}{Character string with a maximum length of 6 characters. Code from a database object. Only one code per iteration.}

\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus'), regionalstatistik.de ('regio'), statistikdaten.bayern.de ('bayern'), landesdatenbank.nrw.de ('nrw'), bildungsmonitoring.de ('bildung') or genesis.sachsen-anhalt.de ('st') database is called. If all databases should be checked, use 'all'. Default option is 'all'.}

\item{area}{Character string. Indicator from which area of the database the results are called. In general, 'all' is the appropriate solution. Default option is 'all'. Not used for 'statistics'.}

\item{sortcriterion}{Character string. Indicator if the output should be sorted by 'code' or 'content'. This is a parameter of the API call itself. The default is 'code'.}

\item{credential_list}{A list containing the credentials for the databases to be accessed. If 'NULL' (default), the function will use the stored credentials from \code{gen_auth_get()}.}

\item{pagelength}{Integer. Maximum length of results or objects (e.g., number of tables). Defaults to 500. Maximum of the databases is 25,000 objects.}

\item{error.ignore}{Boolean. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response. Default option is 'FALSE'. Important note: This does not apply to errors generated by the API call itself (especially server or client errors with status codes 5xx or 4xx). In this case, the function call will error even if error.ignore is set to TRUE.}

\item{verbose}{Boolean. Indicator if the output of the function should include detailed messages and warnings. Default option is 'TRUE'. Set the parameter to 'FALSE' to suppress additional messages and warnings.}

\item{debug}{Boolean. Indicator if the raw output of the API call behind this function should be presented. Default option is 'FALSE'. Set the parameter to 'TRUE' to return raw output.}

\item{...}{Additional parameters for the API call. These parameters are only affecting the call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list with all recalled elements from the API. Always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing with the data. Attributes are added to the data.frame describing the search configuration for the returned output.
}
\description{
Function to search for specific variables. This function gathers metadata, no actual data. It helps to refine data calls.
}
\examples{
\dontrun{
# Find a specific variable "GES" in Genesis
object <- gen_search_vars("GES")
}

}
