% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table.R
\name{gen_table}
\alias{gen_table}
\title{Download Table As Data.Frame}
\usage{
gen_table(
  name,
  database = c("genesis", "zensus", "regio", "bayern", "nrw", "bildung", "st"),
  credential_list = NULL,
  area = c("all", "public", "user"),
  compress = FALSE,
  transpose = FALSE,
  startyear = NULL,
  endyear = NULL,
  regionalvariable = NULL,
  regionalkey = NULL,
  classifyingvariable1 = NULL,
  classifyingkey1 = NULL,
  classifyingvariable2 = NULL,
  classifyingkey2 = NULL,
  classifyingvariable3 = NULL,
  classifyingkey3 = NULL,
  stand = NULL,
  language = Sys.getenv("RESTATIS_LANG"),
  job = FALSE,
  all_character = TRUE,
  ...
)
}
\arguments{
\item{name}{Character string. Name/code of the table. Use of wildcards (\code{*}) is possible.}

\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus'), regionalstatistik.de ('regio'), statistikdaten.bayern.de ('bayern'), landesdatenbank.nrw.de ('nrw'), bildungsmonitoring.de ('bildung') or genesis.sachsen-anhalt.de ('st') database is called.}

\item{credential_list}{A list containing the credentials for the databases to be accessed. If 'NULL' (default), the function will use the stored credentials from \code{gen_auth_get()}.}

\item{area}{Character string. The area in which the table is stored.
Possible values:
\itemize{
\item \code{"public"}: table in the public catalogue
\item \code{"user"}: table in the user's account
\item \code{"all"}: both of the above
}}

\item{compress}{Boolean. Should empty rows and columns be discarded?}

\item{transpose}{Boolean. Reshape the table between \code{"wide"} and \code{"long"} format.}

\item{startyear}{Four-digit integer (valid range 1900 - 2100). Only retrieve data from this year onward. Default 'NULL' can help enable the 'timeslices' parameter (see vignettes for additional parameters).
Note: Starting with {restatis} v0.4.2, no default values for the parameters 'startyear' and 'endyear' will be submitted anymore
(formerly 1900 to 2100 to fetch all available values). Thus, the resulting table might contain a shorter time span than expected.
Instead, we encourage users to explicitly set the desired time span. Doing so, we follow the best practice to save API resources.
The new default of 'NULL' behaves as follows: Each database will always send data, as each table is preconfigured with a standard time
span that varies for each statistic and each table. If the user desires different time spans, she has to specify the values accordingly.}

\item{endyear}{Four-digit integer (valid range 1900 - 2100). Only retrieve data up to this year. Default 'NULL' can help enable the 'timeslices' parameter (see vignettes for additional parameters).
Note: Starting with {restatis} v0.4.2, no default values for the parameters 'startyear' and 'endyear' will be submitted anymore
(formerly 1900 to 2100 to fetch all available values). Thus, the resulting table might contain a shorter time span than expected.
Instead, we encourage users to explicitly set the desired time span. Doing so, we follow the best practice to save API resources.
The new default of 'NULL' behaves as follows: Each database will always send data, as each table is preconfigured with a standard time
span that varies for each statistic and each table. If the user desires different time spans, she has to specify the values accordingly.}

\item{regionalvariable}{Character string. Code of the regional variable whose value
is specified in \code{regionalkey} to filter the results.}

\item{regionalkey}{Character string. One or more regional keys. Multiple values can be
supplied as a character vector or as a single string, with the regional keys
separated by commas. Use of wildcard (\code{*}) allowed.}

\item{classifyingvariable1}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey1} is applied.}

\item{classifyingkey1}{Character string. One or more values of a subject classification
(e.g. \code{"WZ93012"}). Applied to \code{classifyingvariable1}. Multiple keys can be
supplied as a character vector or as a single string, with the keys separated by
commas. Use of wildcard (\code{*}) allowed.}

\item{classifyingvariable2}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey2} is applied.}

\item{classifyingkey2}{Character string. One or more values of a subject classification.
Applied to \code{classifyingvariable2}. Multiple keys can be supplied as a character
vector or as a single string, with the keys separated by commas. Use of wildcard
(\code{*}) allowed.}

\item{classifyingvariable3}{Character string. Code of the subject classification
(SK-Merkmal) to which the selection by means of \code{classifyingkey3} is applied.}

\item{classifyingkey3}{Character string. One or more values of a subject classification.
Applied to \code{classifyingvariable3}. Multiple keys can be supplied as a character
vector or as a single string, with the keys separated by commas. Use of wildcard
(\code{*}) allowed.}

\item{stand}{Character string, format \code{"DD.MM.YYYY"}. Only retrieve data updated
after this date.}

\item{language}{Character string. Search terms, returned messages and data descriptions
in German (\code{"de"}) or English (\code{"en"}).}

\item{job}{Boolean. Indicate whether a job should be created (not available with the
\code{"Zensus"} database). To set \code{job = TRUE}, username and password must be
saved with \code{gen_auth_save()}; using API tokens with \code{job = TRUE} will
result in an error.}

\item{all_character}{Boolean. Should all variables be imported as character variables?
Avoids fuzzy data type conversions if there are leading zeros or other special
characters. Defaults to \code{TRUE}.}

\item{...}{Additional parameters for the API call (see respective API documentation).
A valid specification of these will not be checked by the function, so wrongful
specification may lead to errors.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Download a table with data from one of the supported databases. This is the main function to fetch data, apart from \code{gen_cube}.
}
\examples{
\dontrun{
gen_table("21311-0001", database = "genesis")
}

}
