% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis_variable.R
\name{get_values_from_variables}
\alias{get_values_from_variables}
\title{Get Values From Variables}
\usage{
get_values_from_variables(
  code = NULL,
  sortcriterion = c("code", "content"),
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Destatis-Object. Only one code per iteration.}

\item{sortcriterion}{a string. Indicator if the output should be sorted by 'code' or 'content'. This is a parameter of the Destatis call itself. The default is "code".}

\item{...}{Additional parameter of the Destatis call. These parameters are only affecting the Destatis call itself, no further processing.}
}
\value{
A list with all recalled elements from Destatis. Always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to extract the possible values from a variable from Destatis.
}
\examples{
\dontrun{
# Find the values of the variable "DLAND"
object <- get_values_from_variables(code = "DLAND")
}
}
