% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis_variable.R
\name{get_values_from_variables_from_statistic}
\alias{get_values_from_variables_from_statistic}
\title{Get Values From Variables From Statistic}
\usage{
get_values_from_variables_from_statistic(
  code = NULL,
  detailed.variables = F,
  sortcriterion = c("code", "content"),
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Destatis-Object. Only one code per iteration. "*"-Notations are possibly to be used as a placeholder.}

\item{detailed.variables}{a logical. Indicator if function should return the detailed output of the iteration including all object-related information or only a shortened output including only code and object title. This parameter only affects the details of the variables-related output. The default is FALSE.}

\item{sortcriterion}{a string. Indicator if the output should be sorted by 'code' or 'content'. This is an parameter of the Destatis call itself. The default is "code".}

\item{...}{Additional parameter of the Destatis call. These parameters are only affecting the Destatis call itself for the variables, no further processing or the values-related objects.}
}
\value{
A list with all recalled elements from Destatis. Based on the detailed-parameter it contains more or less information, but always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Get Values From Variables From Statistic
}
\examples{
\dontrun{
# Find the values of variables in a specific statistic with the code "21111" and a detailed description of the variables
object <- get_values_from_variables_from_statistic(code = "21111", detailed.variables = T)
}
}
