% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis_variable.R
\name{get_variables_from_statistic}
\alias{get_variables_from_statistic}
\title{get_variables_from_statistic}
\usage{
get_variables_from_statistic(
  code = NULL,
  detailed = FALSE,
  sortcriterion = c("code", "content"),
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Destatis-Object. Only one code per iteration. "*"-Notations are possibly to be used as a placeholder.}

\item{detailed}{a logical. Indicator if function should return the detailed output of the iteration including all object-related information or only a shortened output including only code and object title. The default is detailed = FALSE.}

\item{sortcriterion}{a string. Indicator if the output should be sorted by 'code' or 'content'. This is a parameter of the Destatis call itself. The default is "code".}

\item{...}{Additional parameter of the Destatis call. These parameters are only affecting the Destatis call itself, no further processing.}
}
\value{
A list with all recalled elements from Destatis. Based on the detailed-parameter it contains more or less information, but always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to generate variables from statistics.
}
\examples{
\dontrun{
# Find the variables of the statistic with the code "12411" with a detailed output
object <- get_variables_from_statistic(code = "12411", detailed = T)
}

}
