% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{insert_and_save_credentials}
\alias{insert_and_save_credentials}
\title{insert_and_save_credentials}
\usage{
insert_and_save_credentials(database)
}
\arguments{
\item{database}{The database to specify credentials for}
}
\description{
insert_and_save_credentials
}
