% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cube.R
\name{is_cube_metadata_header}
\alias{is_cube_metadata_header}
\title{is_cube_metadata_header}
\usage{
is_cube_metadata_header(lines)
}
\arguments{
\item{lines}{Lines to check for header}
}
\description{
is_cube_metadata_header
}
