% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login_check}
\alias{login_check}
\title{Login check}
\usage{
login_check(genesis = getOption("genesis"))
}
\arguments{
\item{genesis}{Which GENESIS database should be used?

One of: "destatis", "regionalstatistik", "bildungsmonitoring", "bayern", "nrw", "sachsen-anhalt".

A default value can also be set to the global option \code{genesis} via
\code{\link[=options]{options()}}, e.g. \code{options(genesis = "destatis")}, so that \code{genesis}
does not have to be specified in the function call.}
}
\description{
Tests if the login with the saved username and password was sucessful
}
\seealso{
\code{\link{set_login_data}}
}
