% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{logincheck_http_error}
\alias{logincheck_http_error}
\title{logincheck_http_error}
\usage{
logincheck_http_error(database, verbose)
}
\arguments{
\item{database}{The user input to 'gen_logincheck'}

\item{verbose}{Boolean. Should the function message in case of success?}
}
\value{
Informative error/warning messages + invisibly TRUE/FALSE
}
\description{
logincheck_http_error
}
