% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{logincheck_stop_or_warn}
\alias{logincheck_stop_or_warn}
\title{logincheck_stop_or_warn}
\usage{
logincheck_stop_or_warn(response, error, verbose, database)
}
\arguments{
\item{response}{A HTTP response object}

\item{error}{Boolean. Should the function warn or throw an error?}

\item{verbose}{Boolean. Should the function message in case of success?}

\item{database}{The database that the check should be run for}
}
\value{
In case of failure warns or errors. Invisibly returns TRUE (success) or FALSE (failure)
}
\description{
logincheck_stop_or_warn
}
