% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_statistic}
\alias{metadata_statistic}
\alias{metadata_table}
\alias{metadata_value}
\alias{metadata_variable}
\title{Get metadata}
\usage{
metadata_statistic(name, area = c("free", "user", "all"), language = "en")

metadata_table(name, area = c("free", "user", "all"), language = "en")

metadata_value(name, area = c("free", "user", "all"), language = "en")

metadata_variable(name, area = c("free", "user", "all"), language = "en")
}
\arguments{
\item{name}{Code of the variable (e.g. retrieved using \link{catalogue_variables})}

\item{area}{Area in which the object is stored:
\itemize{
\item "free" = objects in the public catalogue
\item "user" = objects saved to "Meine Tabellen"
}}

\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}
}
\description{
Retrieves metadata for a statistic, table, variable or value
}
\examples{
\dontrun{
metadata_statistic("12411")
metadata_table("12411-0001")
metadata_variable("KREISE")
metadata_value("VV101")
}
}
