% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis_modifydata.R
\name{modified_data}
\alias{modified_data}
\title{modified_data}
\usage{
modified_data(
  code = "",
  type = c("all", "tables", "statistics", "statisticsUpdates"),
  date = c("now", "week_before", "month_before", "year_before"),
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Genesis object. Only one code per iteration. "*" notations are possible.}

\item{type}{a string. Specific Genesis object type: 'tables', 'statistics', and 'statisticsUpdates'. All three can be accessed through "all", which is the default.}

\item{date}{a string. Specific date that is used as the last update or upload time in Genesis to include a Genesis object in return. Default option is 'now', which uses the current date of your system. Alternative options are 'week_before', using the current date of your system minus 7 days, 'month_before', using the current date of your system minus 4 weeks, and 'year_before', using the current date of your system minus 52 weeks. Additionally, it is possible to fill in a specific date of format 'DD.MM.YYYY'.}

\item{...}{Additional parameters for the Genesis call. These parameters are only affecting the Genesis call itself, not the further processing.}
}
\value{
A list with all recalled elements from Genesis. Always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing with the data. Attributes are added to the data.frame describing the search configuration for the returned output.
}
\description{
Function to check for updates or new objects in Genesis based on a specific date.
}
\examples{
\dontrun{
# Find Genesis objects which were modified last_week
object <- modified_data(date = "week_before")

# Find tables that were new from 31.03.2020
object <- modified_data(type = "tables", date = "31.03.2020")

# Find Destatis-Objects related to the topic "Bevölkerung" (Code: '12*') which were new today
object <- modified_data(code = "12*")
}

}
