% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{param_check_regionalkey}
\alias{param_check_regionalkey}
\title{param_check_regionalkey}
\usage{
param_check_regionalkey(regionalkey)
}
\arguments{
\item{regionalkey}{Regional key}
}
\description{
param_check_regionalkey
}
