% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis-api.R
\name{print.genesis_api}
\alias{print.genesis_api}
\title{Printing parsed GENESIS API request}
\usage{
\method{print}{genesis_api}(x, ...)
}
\arguments{
\item{x}{Object of class \code{genesis_api}}

\item{...}{Optional arguments to print methods}
}
\description{
Printing parsed GENESIS API request
}
