% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis-api.R
\name{print.genesis_find}
\alias{print.genesis_find}
\title{Printing parsed GENESIS API request for \code{Find} methods}
\usage{
\method{print}{genesis_find}(x, ...)
}
\arguments{
\item{x}{Object of class \code{genesis_find}}

\item{...}{Optional arguments to print methods}
}
\description{
Printing parsed GENESIS API request for \code{Find} methods
}
