% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis-api.R
\name{print.genesis_helloworld}
\alias{print.genesis_helloworld}
\title{Printing parsed GENESIS API request for \code{HelloWorld} methods}
\usage{
\method{print}{genesis_helloworld}(x, ...)
}
\arguments{
\item{x}{Object of class \code{genesis_helloworld}}

\item{...}{Optional arguments to print methods}
}
\description{
Printing parsed GENESIS API request for \code{HelloWorld} methods
}
