% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cube.R
\name{read_cube_data_lines}
\alias{read_cube_data_lines}
\title{read_cube_data_lines}
\usage{
read_cube_data_lines(lines, col_names)
}
\arguments{
\item{lines}{Lines to read data from}

\item{col_names}{Specify column names}
}
\description{
read_cube_data_lines
}
