% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cube.R
\name{read_cube_metadata_header}
\alias{read_cube_metadata_header}
\title{read_cube_metadata_header}
\usage{
read_cube_metadata_header(line, rename_dups = TRUE)
}
\arguments{
\item{line}{Line to read}

\item{rename_dups}{Rename duplicates?}
}
\description{
read_cube_metadata_header
}
