% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cube.R
\name{rename_cube_data_columns}
\alias{rename_cube_data_columns}
\title{rename_cube_data_columns}
\usage{
rename_cube_data_columns(cube)
}
\arguments{
\item{cube}{A cube object to rename the columns in}
}
\description{
rename_cube_data_columns
}
