% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{return_table_object}
\alias{return_table_object}
\title{return_table_object}
\usage{
return_table_object(response, response_type, language, all_character)
}
\arguments{
\item{response}{Response object}

\item{response_type}{Response type}

\item{language}{Language locale}

\item{all_character}{Read all variables as character?}
}
\description{
return_table_object
}
