% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{rev_database_function}
\alias{rev_database_function}
\title{rev_database_function}
\usage{
rev_database_function(input)
}
\arguments{
\item{input}{Input to test for database name}
}
\description{
rev_database_function
}
