% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{search_tables}
\alias{search_tables}
\alias{search_statistics}
\alias{search_variables}
\title{Find tables, statistics, variables, data cubes, or time series}
\usage{
search_tables(term, pagelength = 100, language = "en")

search_statistics(term, pagelength = 100, language = "en")

search_variables(term, pagelength = 100, language = "en")
}
\arguments{
\item{term}{Term to search for}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves lists of objects for a search term
(tables, statistics, characteristics, data cubes or time series)
}
\examples{
\dontrun{
search_statistics("Abfall", pagelength = 6, language = "de")

search_variables("vote")
}
}
\seealso{
catalogue_terms
}
