% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis_variable.R
\name{search_variables}
\alias{search_variables}
\title{Search For Variable In Destatis}
\usage{
search_variables(code = NULL, sortcriterion = c("code", "content"), ...)
}
\arguments{
\item{code}{a string with a maximum length of 6. Code from a Destatis-Object. Only one code per iteration. "*"-Notations are possibly to be used as a placeholder.}

\item{sortcriterion}{a string. Indicator if the output should be sorted by 'code' or 'content'. This is a parameter of the Destatis call itself. The default is "code".}

\item{...}{Additional parameter of the Destatis call. These parameters are only affecting the Destatis call itself, no further processing.}
}
\value{
A list with all recalled elements from Destatis. Always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to search for specific variables in Destatis databank.
}
\examples{
\dontrun{
# Find a specific variable "GES" in Destatis
object <- search_variables("GES")
}
}
