% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{set_credentials_auth}
\alias{set_credentials_auth}
\title{set_credentials_auth}
\usage{
set_credentials_auth(path, sys_env, ui_menu_database, use_token)
}
\arguments{
\item{path}{Path for the .rds file}

\item{sys_env}{System environment variable name for the key}

\item{ui_menu_database}{The database for the auth request ('GENESIS' or 'Zensus 2022')}

\item{use_token}{Boolean. Do you want to (if possible) set an API token instead of password + username? Defaults to FALSE.}
}
\description{
set_credentials_auth
}
