% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{spezifisch_create}
\alias{spezifisch_create}
\title{spezifisch_create}
\usage{
spezifisch_create(x)
}
\arguments{
\item{x}{Element to extract $Spezifisch from}
}
\description{
spezifisch_create
}
