% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_authentication.R
\name{test_database_function}
\alias{test_database_function}
\title{test_database_function}
\usage{
test_database_function(input, credential_list, error.input, text)
}
\arguments{
\item{input}{Input to test for database name}

\item{credential_list}{A list containing the credentials for the databases to be accessed. If 'NULL' (default), the function will use the stored credentials from \code{gen_auth_get()}.}

\item{error.input}{Indicator error.ignore}

\item{text}{Indicator verbose}
}
\description{
test_database_function
}
