% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{test_if_error}
\alias{test_if_error}
\title{test_if_error}
\usage{
test_if_error(input, para, verbose = NULL)
}
\arguments{
\item{input}{Response object}

\item{para}{Parameter TRUE/FALSE}

\item{verbose}{Verbose TRUE/FALSE}
}
\description{
test_if_error
}
