% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api_calls.R
\name{test_if_error_light}
\alias{test_if_error_light}
\title{test_if_error_light}
\usage{
test_if_error_light(input, verbose = NULL)
}
\arguments{
\item{input}{Response object}

\item{verbose}{Verbose TRUE/FALSE}
}
\description{
test_if_error_light
}
