% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{test_if_error_variables}
\alias{test_if_error_variables}
\title{test_if_error_variables}
\usage{
test_if_error_variables(input, para)
}
\arguments{
\item{input}{Response object}

\item{para}{Parameter TRUE/FALSE}
}
\description{
test_if_error_variables
}
