% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{titel_search}
\alias{titel_search}
\title{titel_search}
\usage{
titel_search(x, term, text)
}
\arguments{
\item{x}{Element to extract $Content from}

\item{term}{Search term}

\item{text}{Indicator verbose}
}
\description{
titel_search
}
