% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{title_search}
\alias{title_search}
\title{title_search}
\usage{
title_search(x, term, text)
}
\arguments{
\item{x}{Element to extract $Content from}

\item{term}{Search term}

\item{text}{Indicator verbose}
}
\description{
title_search
}
