% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Restatis_xytovariable.R
\name{xy_to_variable}
\alias{xy_to_variable}
\title{Get Objects Related To Variables}
\usage{
xy_to_variable(
  code = NULL,
  category = c("tables", "statistics", "cubes"),
  detailed = F,
  sortcriterion = c("code", "content"),
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Destatis-Object. Only one code per iteration.}

\item{category}{a string. Specific Destatis-Object-types: 'tables', 'statistics', and 'cubes'. All three together are possible and the default option.}

\item{detailed}{a logical. Indicator if function should return the detailed output of the iteration including all object-related information or only a shortened output including only code and object title. The default is detailed = FALSE.}

\item{sortcriterion}{a string. Indicator if the output should be sorted by 'code' or 'content'. This is a parameter of the Destatis call itself. The default is "code".}

\item{...}{Additional parameter of the Destatis call. These parameters are only affecting the Destatis call itself, no further processing.}
}
\value{
A list with all recalled elements from Destatis. Based on the detailed-parameter it contains more or less information, but always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to find objects related to a variable in Destatis.
}
\examples{
\dontrun{
# Find Tables for Variable "Kreise" and return detailed output
object <- xy_to_variable(code = "Kreise", category = "tables", detailed = T)

#' # Find everything for Variable "GES"
object <- xy_to_variable(code = "GES")
# Default of detailed-parameter is FALSE, and default of the category-parameter is to include all object types.
}
}
