% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue_qualitysigns}
\alias{catalogue_qualitysigns}
\title{Quality indicator catalogue}
\usage{
catalogue_qualitysigns(language = "en")
}
\arguments{
\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieve list of quality indicators
}
