% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue-variables.R
\name{catalogue_variables}
\alias{catalogue_variables}
\title{Variable catalogue}
\usage{
catalogue_variables(
  selection,
  area = c("free", "user", "all"),
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  pagelength = 100,
  language = "en"
)
}
\arguments{
\item{selection}{String to filter for the code or content of the object(s) to
be returned. Use of wildcard (*) possible.}

\item{area}{Area in which the object is stored:
\itemize{
\item "free" = objects in the public catalogue
\item "user" = objects saved to "Meine Tabellen"
}}

\item{searchcriterion}{Does \code{selection} refer to the code or the content of the object(s)?}

\item{sortcriterion}{Should the results be sorted by the code or the content of the object(s)?}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves list of variables according to selection
}
\examples{
\dontrun{
catalogue_variables("FA*")
}
}
