% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{check_function_input}
\alias{check_function_input}
\title{check_function_input}
\usage{
check_function_input(
  code = NULL,
  term = NULL,
  sortcriterion = NULL,
  category = NULL,
  detailed = NULL,
  type = NULL,
  date = NULL,
  similarity = NULL,
  error.ignore = NULL,
  ordering = NULL,
  pagelength = NULL,
  database = NULL,
  area = NULL,
  caller = NULL,
  verbose = NULL,
  debug = NULL,
  raw = NULL,
)
}
\arguments{
\item{code}{Parameter to be checked}

\item{term}{Parameter to be checked}

\item{sortcriterion}{Parameter to be checked}

\item{category}{Parameter to be checked}

\item{detailed}{Parameter to be checked}

\item{type}{Parameter to be checked}

\item{date}{Parameter to be checked}

\item{similarity}{Parameter to be checked}

\item{error.ignore}{Parameter to be checked}

\item{ordering}{Parameter to be checked}

\item{pagelength}{Parameter to be checked}

\item{database}{Parameter to be checked}

\item{area}{Parameter to be checked}

\item{caller}{Parameter to be checked}

\item{verbose}{Parameter to be checked}

\item{debug}{Parameter to be checked}

\item{raw}{Parameter to be checked}
}
\description{
check_function_input
}
