% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api.R
\name{.gen_api_core}
\alias{.gen_api_core}
\title{Basic API request function}
\usage{
.gen_api_core(endpoint, database, credential_list = NULL, ...)
}
\arguments{
\item{endpoint}{Character string. The endpoint of the API that is to be queried.}

\item{database}{The database the query should be sent to.}

\item{credential_list}{A named list including username and password for the database(s) to be queried. If NULL (default), the function will try to get credentials from the environment variables.}

\item{...}{Further parameters passed on to the final API call.}
}
\description{
(Uncached) Low-level function to interact with the one of the APIs
}
