% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api.R
\name{.gen_api_core}
\alias{.gen_api_core}
\title{.gen_api_core}
\usage{
.gen_api_core(endpoint, database, ...)
}
\arguments{
\item{endpoint}{Character string. The endpoint of the API that is to be queried.}

\item{database}{The database the query should be sent to.}

\item{...}{Further parameters passed on to the final API call.}
}
\description{
Low-level function to interact with the one of the APIs
}
