% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_tables}
\alias{find_tables}
\alias{find_statistics}
\alias{find_variables}
\alias{find_cubes}
\alias{find_timeseries}
\title{Find tables, statistics, variables, data cubes, or time series}
\usage{
find_tables(term, pagelength = 100, language = "en")

find_statistics(term, pagelength = 100, language = "en")

find_variables(term, pagelength = 100, language = "en")

find_cubes(term, pagelength = 100, language = "en")

find_timeseries(term, pagelength = 100, language = "en")
}
\arguments{
\item{term}{Term to search for}

\item{pagelength}{Maximum number of results delivered}

\item{language}{Search terms, returned messages and data descriptions in German ("de") or English ("en")?}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves lists of objects for a search term
(tables, statistics, characteristics, data cubes or time series)
}
\examples{
\dontrun{
find_statistics("Abfall", pagelength = 6, language = "de")

find_variables("vote")
}
}
