% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{gen_cube}
\alias{gen_cube}
\title{Get a data cube}
\usage{
gen_cube(name, ...)
}
\arguments{
\item{name}{Name of the data cube}

\item{...}{TODO}
}
\description{
Get a data cube
}
\examples{
\dontrun{
gen_cube("47414BJ002")
}
}
