% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_list_jobs.R
\name{gen_list_jobs}
\alias{gen_list_jobs}
\title{gen_list_jobs: Explore current jobs of your user account}
\usage{
gen_list_jobs(
  selection = NULL,
  sortcriterion = c("content", "time", "status"),
  ...
)
}
\arguments{
\item{selection}{Filter the list of jobs for matching codes.}

\item{sortcriterion}{Allows to sort the resulting list of jobs by their Code ("content"), the time of completion ("time") or status ("status")}

\item{...}{Additional parameters for the Genesis API call. These parameters are only affecting the Genesis call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list of all current jobs connected to the given user.
}
\description{
Function to list all current jobs connected to the given user.
}
