% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_jobs.R, R/gen_list_jobs.R
\name{gen_list_jobs}
\alias{gen_list_jobs}
\title{gen_list_jobs}
\usage{
gen_list_jobs(
  selection = NULL,
  sortcriterion = c("content", "time", "status"),
  ...
)

gen_list_jobs(
  selection = NULL,
  sortcriterion = c("content", "time", "status"),
  ...
)
}
\arguments{
\item{selection}{Filter the list of jobs for matching codes.}

\item{sortcriterion}{Allows to sort the resulting list of jobs by their Code ("content"), the time of completion ("time") or status ("status")}

\item{...}{Additional parameter of the GENESIS API call. These parameters are only affecting the GENESIS API call itself, no further processing.}

\item{database}{Character string. Indicator if 'genesis' or 'regionalstatistik.de' database is called. Default option is 'genesis'.}

\item{flat}{Boolean. Should the function return a list with jobs and metadata ('FALSE') or just a flat data.frame ('TRUE')? Defaults to FALSE.}
}
\value{
A list or data.frame (see parameter 'flat') of all current jobs of the user.

A list of all current jobs connected to the given user.
}
\description{
Function to list all current jobs connected to the given user in the GENESIS or regionalstatistik.de database. Important note: For this function it is also possible to use \code{searchcriterion} parameter and \code{selection} parameter, making it possible to filter the job list based on 'type','time','status' or 'code'. For more details see \code{vignette("additional_parameter")}.

Function to list all current jobs connected to the given user.
}
\examples{
\dontrun{
gen_list_jobs("regio", flat = TRUE)
}

}
