% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_logincheck.R
\name{gen_logincheck}
\alias{gen_logincheck}
\title{gen_logincheck}
\usage{
gen_logincheck(database = "all", verbose = TRUE, ...)
}
\arguments{
\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus'), regionalstatistik.de ('regio'), statistikdaten.bayern.de ('bayern'), landesdatenbank.nrw.de ('nrw'), bildungsmonitoring.de ('bildung') or genesis.sachsen-anhalt.de ('sa') database is called. If all databases should be checked, use 'all'. Default option is 'all'.}

\item{verbose}{Boolean. In case of success, should a message be printed? Defaults to 'TRUE'.}

\item{...}{Additional parameters for the API call. These parameters are only affecting the call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
Leads to an informative error message if the login check failed and returns FALSE invisibly. Invisibly returns TRUE otherwise.
}
\description{
Function to check if a login is possible for a certain database.
}
\examples{
\dontrun{
gen_logincheck("zensus")
}

}
