% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_objects2var.R
\name{gen_objects2var}
\alias{gen_objects2var}
\title{gen_objects2var: Get Objects Related To Variable}
\usage{
gen_objects2var(
  code = NULL,
  category = c("tables", "statistics", "cubes"),
  detailed = FALSE,
  error.ignore = FALSE,
  ...
)
}
\arguments{
\item{code}{a string with a maximum length of 15 characters. Code from a Genesis-Object. Only one code per iteration.}

\item{category}{a string. Specific object-types: 'tables', 'statistics', and 'cubes'. All three together are possible and the default option.}

\item{detailed}{a logical. Indicator if function should return the detailed output of the iteration including all object-related information or only a shortened output including only code and object title. The default is detailed = FALSE.}

\item{error.ignore}{a logical. Indicator if the function should stop if an error occurs or no object for the request is found or if it should produce a token as response.}

\item{...}{Additional parameters for the Genesis API call. These parameters are only affecting the Genesis call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list with all recalled elements from Genesis. Based on the detailed-parameter it contains more or less information, but always includes the code of the object, the title, and the type of the object. This is done to facilitate further processing of the data. Attributes are added to the dataframe describing the search configuration for the returned output.
}
\description{
Function to find objects related to a variable in Genesis.
}
\examples{
\dontrun{
# Find Tables for Variable "Kreise" and return detailed output
object <- gen_objects2var(code = "Kreise", category = "tables", detailed = T)

# Find everything for Variable "GES"
object <- gen_objects2var(code = "GES")
# Default of detailed-parameter is FALSE, and default of the
# category-parameter is to include all object types.
}

}
