% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_qualitysigns.R
\name{gen_signs}
\alias{gen_signs}
\title{gen_signs}
\usage{
gen_signs(database = c("all", "genesis", "zensus", "regio"), ...)
}
\arguments{
\item{database}{Character string. Indicator if the GENESIS ('genesis'), Zensus 2022 ('zensus') or regionalstatistik.de ('regio') database is called. Default option is 'all'.}

\item{...}{Additional parameters for the API call. These parameters are only affecting the call itself, no further processing. For more details see \code{vignette("additional_parameter")}.}
}
\value{
A list of all current used special signs.
}
\description{
Function to list all currently used special signs (e.g., 0, *, X, (), p, ...) and their meaning in GENESIS, Zensus 2022 and/or regionalstatistik.de.
}
