% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{gen_table}
\alias{gen_table}
\title{gen_table}
\usage{
gen_table(name, ...)
}
\arguments{
\item{name}{Name of the table}

\item{...}{TODO}
}
\description{
Download a table with data from Genesis
}
\examples{
\dontrun{
gen_table("21311-0001")
}

}
