% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_table}
\alias{get_table}
\title{Get a data table}
\usage{
get_table(
  name,
  area = c("free", "user", "all"),
  compress = FALSE,
  startyear = 1900,
  endyear = 2100,
  timeslices = NULL,
  regionalvariable = NULL,
  regionalkey = NULL,
  job = NULL,
  stand = NULL,
  language = getOption("genesis_language"),
  genesis = getOption("genesis")
)
}
\arguments{
\item{name}{Code of the variable (e.g. retrieved using \link{catalogue_variables})}

\item{area}{Area in which the object is stored:
\itemize{
\item "free" = objects in the public catalogue
\item "user" = objects saved to "Meine Tabellen"
}}

\item{compress}{Should empty rows and columns be discarded?}

\item{startyear, endyear}{Only retrieve data between these years}

\item{timeslices}{Number of latest time slices to retrieve --
independent of/cumulative to \code{startyear} and \code{endyear}}

\item{regionalvariable}{Code of the regional variable whose value is
specified in \code{regionalkey} to filter the results}

\item{regionalkey}{One or more regional keys. Multiple values can be supplied
as a character vector or as a single string, with the regional keys
separated by commas. Use of wildcard (*) possible.}

\item{job}{Should a job be created if the table cannot be created immediately?}

\item{stand}{Only retrieve data updated after this date (dd.mm.yyyy)}

\item{language}{Search terms, returned messages and data descriptions
in German ("de") or English ("en")?

A default value can also be set to the global option \code{genesis_language} via
\code{\link[=options]{options()}}, e.g. \code{options(genesis_language = "de")}.}

\item{genesis}{Which GENESIS database should be used?

One of: "destatis", "regionalstatistik", "bildungsmonitoring", "bayern", "nrw", "sachsen-anhalt".

A default value can also be set to the global option \code{genesis} via
\code{\link[=options]{options()}}, e.g. \code{options(genesis = "destatis")}, so that \code{genesis}
does not have to be specified in the function call.}
}
\value{
A \code{data.frame} (or \code{tbl_df} if tibble package is installed)
}
\description{
Retrieves a data table based on the name of the table
}
\examples{
\dontrun{
options(genesis = "destatis")
get_table("12411-0001", startyear = 1960, endyear = 1969)
# Filter results by region
# ("KREISE": administrative districts; "01001": Flensburg)
get_table("12411-0015", regionalvariable = "KREISE", regionalkey = "01001")

options(genesis = "regionalstatistik")
# Get large table by creating a background processing job
get_table("12711-01-03-4", job = TRUE)
# List jobs you've created
catalogue_jobs()
# After the job is finished, you can retrieve the table with `get_table()`
# with `name` set to the name of the job.
}
}
