% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis-package.R
\docType{package}
\name{restatis-package}
\alias{restatis}
\alias{restatis-package}
\title{restatis: R Wrapper to Access a Wide Range of Germany's Federal Statistical System Databases Based on the GENESIS Web Service RESTful API of the German Federal Statistical Office (Statistisches Bundesamt/Destatis)}
\description{
A RESTful API wrapper for accessing the GENESIS database of the German Federal Statistical Office (Destatis) as well as its Census Database and the database of Germany's regional statistics. Supports data search functions, credential management, result caching, and handling remote background jobs for large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://correlaid.github.io/restatis/}
  \item \url{https://github.com/CorrelAid/restatis}
  \item Report bugs at \url{https://github.com/CorrelAid/restatis/issues}
}

}
\author{
\strong{Maintainer}: Yannik Buhl \email{ybuhl@posteo.de}

Authors:
\itemize{
  \item Zoran Kovacevic (\href{https://orcid.org/0009-0002-0156-0862}{ORCID})
  \item Dorian Le Jeune
  \item Long Nguyen \email{long.nguyen@uni-bielefeld.de} (\href{https://orcid.org/0000-0001-8878-7386}{ORCID})
  \item Johannes Ritter \email{ritter.johannes@gmail.com}
}

}
\keyword{internal}
