% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatis-package.R
\docType{package}
\name{restatis-package}
\alias{restatis}
\alias{restatis-package}
\title{restatis: Web API Client for the German Federal Statistical Office Database}
\description{
A 'RESTful' API wrapper for accessing the 'GENESIS' database of the German Federal Statistical Office (Destatis) \url{https://www-genesis.destatis.de/}. Also supports data search functions, credential management, result caching, and handling remote background jobs for large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://correlaid.github.io/restatis/}
  \item \url{https://github.com/CorrelAid/restatis}
  \item Report bugs at \url{https://github.com/CorrelAid/restatis/issues}
}

}
\author{
\strong{Maintainer}: Long Nguyen \email{long.nguyen@uni-bielefeld.de} (\href{https://orcid.org/0000-0001-8878-7386}{ORCID})

Authors:
\itemize{
  \item Yannik Buhl \email{yannik.buhl@posteo.de}
  \item Zoran Kovacevic \email{zoran.kovacevic@gess.ethz.ch} (\href{https://orcid.org/0009-0002-0156-0862}{ORCID})
  \item Dorian Le Jeune \email{d.le-jeune@gmx.de}
  \item Johannes Ritter \email{ritter.johannes@gmail.com}
}

}
\keyword{internal}
