% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{test_database_function}
\alias{test_database_function}
\title{test_database_function}
\usage{
test_database_function(input)
}
\arguments{
\item{input}{Input to test for database name}
}
\description{
test_database_function
}
