% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataprocessing.R
\name{test_database_function}
\alias{test_database_function}
\title{test_database_function}
\usage{
test_database_function(input, error.input, text)
}
\arguments{
\item{input}{Input to test for database name}

\item{error.input}{Indicator error.ignore}

\item{text}{Indicator verbose}
}
\description{
test_database_function
}
